@@ -23,40 +23,37 @@
     "@angular/router": "^5.2.0",
     "body-parser": "^1.18.2",
     "cookie-parser": "^1.4.3",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.5",
     "dotenv": "^5.0.1",
     "exif": "^0.6.0",
     "express": "^4.16.3",
     "helmet": "^3.12.0",
     "moment": "^2.22.0",
-    "mongoose": "^5.0.12",
+    "mongoose": "^5.0.14",
     "multer": "^1.3.0",
     "passport": "^0.4.0",
     "path": "^0.12.7",
-    "rxjs": "^5.5.6",
+    "rxjs": "^5.5.10",
     "sharp": "^0.20.1",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
+    "@angular/cli": "~6.0.0-rc.4",
     "@angular/compiler-cli": "^5.2.0",
     "@angular/language-service": "^5.2.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "~9.6.4",
     "codelyzer": "^4.0.1",
     "eslint": "^4.19.1",
     "eslint-config-google": "^0.9.1",
-    "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.8.1"
   }
 }
